Make this logic a little clearer.
